+++ packages/natives/native/index.d.ts
@@ -93,7 +93,7 @@ export declare class TtyWriter {
 
 export declare function __ompInstallTokioRuntime(): void
 
-export declare function __piNativesV18_0_6(): void
+export declare function __piNativesV18_1_6(): void
 
 export declare function astEdit(options: AstReplaceOptions): Promise<AstReplaceResult>